filtering null txes as well as batching firehose writes to overcome 413s
